requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
awsebcli==3.15.2
Babel==2.7.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
blinker==1.4
cement==2.8.2
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.4.1
cryptography==2.7
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
future==0.16.0
idna==2.8
imagesize==1.1.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-core==4.5.0
keyring==18.0.0
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
mysqlclient==1.4.2.post1
nbconvert==5.5.0
nbformat==4.4.0
numpydoc==0.9.1
packaging==19.0
pandocfilters==1.4.2
parso==0.5.0
pathspec==0.5.9
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
psutil==5.6.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.8.0
requests==2.20.1
rope==0.14.0
semantic-version==2.5.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
termcolor==1.1.0
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
WTForms==2.2.1
